chore: rename store to stores
